list slicing
Watson gives Sherlock a list of N numbers. Then he asks him to determine if
there exists an element in the list such that the sum of the elements on its
left is equal to the sum of the elements on its right. If there are no elements
to the right/left, then the sum is considered to be zero.